refactor(sealedCopy): move isParameterAssignableFromProperty to its own file (#162)
Used by both SealedCopyLeaf.kt and SealedCopyViaValidation.kt.
